#49 - refactor and cs:fix
