+++ src/Http/Routing/WebRouting.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Blumilk\Meetup\Core\Http\Routing;
 
 use Blumilk\Meetup\Core\Http\Controllers\Auth\LoginController;
+use Blumilk\Meetup\Core\Http\Controllers\Auth\PasswordResetController;
 use Blumilk\Meetup\Core\Http\Controllers\Auth\RegisterController;
 use Blumilk\Meetup\Core\Http\Controllers\Auth\SocialiteController;
 use Blumilk\Meetup\Core\Http\Controllers\ContactController;
 use Blumilk\Meetup\Core\Http\Controllers\MeetupController;
 use Blumilk\Meetup\Core\Http\Controllers\OrganizationController;
-use Blumilk\Meetup\Core\Http\Controllers\PasswordResetController;
 use Blumilk\Meetup\Core\Http\Controllers\SpeakersController;
 use Blumilk\Meetup\Core\Http\Controllers\StaticController;
